// src/documents/enums/document-type.enum.ts
export enum DocumentType {
  IDENTITY = 'IDENTITY',
  CONTRACT = 'CONTRACT',
  PROOF_OF_PAYMENT = 'PROOF_OF_PAYMENT',
  RECEIPT = 'RECEIPT',
  INVOICE = 'INVOICE',
  PAYMENT_SLIP = 'PAYMENT_SLIP',
  PASSPORT = 'PASSPORT',
  ID_CARD = 'ID_CARD',
  DRIVER_LICENSE = 'DRIVER_LICENSE',
  VEHICLE_REGISTRATION = 'VEHICLE_REGISTRATION',
}